eventList with infinite Scroll

+++ root/backend/routes/api/event.js
@@ -5,17 +5,37 @@ const router = express.Router();
 const mongoose = require("mongoose");
 // Models
 const Event = require("../../models/Event");
-const User=require("../../models/User")
-Joi.objectId = require('joi-objectid')(Joi);
+const User = require("../../models/User");
+Joi.objectId = require("joi-objectid")(Joi);
 
-// make sure to notify about object id  and pretty method 
+// make sure to notify about object id  and pretty method
 
 ///////////CRUDZZZZZZZ\\\\\\\\\\\\
 // Read all Events
 router.get("/", async(req, res) => {
   const events = await Event.find().populate('organizerId').populate('eventRequestId')
-    
   res.json({ data: events }); // make sure of pretty
+});
+//----------------------------------------------\\
+
+// Read Events in specific range
+router.get("/withRange/:limit/:offset", async (req, res) => {
+  const schema = {
+    limit: Joi.required(),
+    offset: Joi.required()
+  };
+
+  const result = Joi.validate(req.params, schema);
+
+  if (result.error)
+    return res.status(400).send({ error: result.error.details[0].message });
+
+  const limit = parseInt(req.params.limit, 10);
+  const offset = parseInt(req.params.offset, 10);
+  const events = await Event.find()
+    .skip(offset)
+    .limit(limit);
+  res.json({ data: events });
 });
 //----------------------------------------------\\
 
@@ -78,52 +98,56 @@ res.json({ data: events });
 // });
 //----------------------------------------------------\\
 
-
 router.post("/", async (req, res) => {
   const schema = {
-    location: Joi.object().keys(
-        {  
-            city :Joi.string(),
-            Street :Joi.string() , 
-            Area :Joi.string() 
-        }
-    ).required(), 
-    description: Joi.string().min(30).required(),
+    location: Joi.object()
+      .keys({
+        city: Joi.string(),
+        Street: Joi.string(),
+        Area: Joi.string()
+      })
+      .required(),
+    description: Joi.string()
+      .min(30)
+      .required(),
     type: Joi.string().required(),
     registrationPrice: Joi.number().required(),
     numberOfSpaces: Joi.number().required(),
-    speakers: Joi.array().items(Joi.string()).required(),
-    topics: Joi.array().items(Joi.string()).required(),
+    speakers: Joi.array()
+      .items(Joi.string())
+      .required(),
+    topics: Joi.array()
+      .items(Joi.string())
+      .required(),
     dateTime: Joi.date().required(),
     organizerId: Joi.objectId().required(),
     numberOfRegisterations: Joi.number().required(),
-    eventRequestId: Joi.objectId().required() , 
-    rate : Joi.number().min(0).max(5) 
+    eventRequestId: Joi.objectId().required(),
+    rate: Joi.number()
+      .min(0)
+      .max(5)
   };
 
-  const result =  Joi.validate(req.body, schema);
+  const result = Joi.validate(req.body, schema);
 
   if (result.error)
     return res.status(400).send({ error: result.error.details[0].message });
-  
-  const newEvent = await Event.create(req.body)
+
+  const newEvent = await Event.create(req.body);
   return res.json({ data: newEvent });
 });
 //----------------------------------------\\
 
 // Update Event
-router.put("/:id",async (req, res) => {
+router.put("/:id", async (req, res) => {
   const requestedId = req.params.id;
 
-
   const schema = {
-    location: Joi.object().keys(
-        { 
-            city :Joi.string(),
-            Street :Joi.string() , 
-            Area :Joi.string() , 
-        }
-    ), 
+    location: Joi.object().keys({
+      city: Joi.string(),
+      Street: Joi.string(),
+      Area: Joi.string()
+    }),
     description: Joi.string().min(30),
     type: Joi.string(),
     registrationPrice: Joi.number(),
@@ -133,28 +157,35 @@ router.put("/:id",async (req, res) => {
     dateTime: Joi.date(),
     organizerId: Joi.objectId(),
     numberOfRegisterations: Joi.number(),
-    eventRequestId: Joi.objectId(), 
-    rate : Joi.number().min(0).max(5) 
+    eventRequestId: Joi.objectId(),
+    rate: Joi.number()
+      .min(0)
+      .max(5)
   };
   const result = Joi.validate(req.body, schema);
 
   if (result.error)
     return res.status(400).send({ error: result.error.details[0].message });
-  const updatedEvent = await Event.updateOne({'_id':requestedId},req.body)//test $set as it's not in the lab
+  const updatedEvent = await Event.updateOne({ _id: requestedId }, req.body); //test $set as it's not in the lab
 
-  res.send({data : updatedEvent});
+  res.send({ data: updatedEvent });
 });
 //-----------------------------------\\
 
 // Delete a Event
-router.delete("/:id", async(req, res) => {
+router.delete("/:id", async (req, res) => {
   const requestedId = req.params.id;
-  
   const event = Event.findOne({requestedId})
   if(!event) return res.status(400).send({error: 'The Event you are tryinig to edit does not exist'})
 
-  const deletedEvent = await Event.findByIdAndRemove(requestedId)
-  res.send({'you have deleted ' : deletedEvent})
+  const event = Event.findOne({ requestedId });
+  if (!event)
+    return res
+      .status(404)
+      .send({ error: "The Event you are tryinig to edit does not exist" });
+
+  const deletedEvent = await Event.findByIdAndRemove(requestedId);
+  res.send({ "you have deleted ": deletedEvent });
 });
 //---------------------------------\\
